Persist frame placement through sessions
The problem was caused by two things:

1) I had set Fishbringer:SetUserPlaced(false) due to unexpected
placement behavior during development. Which is caused by ...

2) I swapped around the different events used to load the addon, as
some were no longer emitted. SetUserPlaced(true) and frames to have
coordinates stored using it needs to be placed before the PLAYER_LOGIN
event.

This change will extract the database initialization (for alignments,
hidden elements, etc) and the frame initialization to be triggered for
the ADDON_LOADED event.